+++ app/configuration/extensions/database.py
@@ -1,8 +1,13 @@
 # -*- coding: utf-8 -*-
-from app.configuration.extensions import db
+from flask.ext.sqlalchemy import SQLAlchemy
 
 
-def initDatabase(app):
+db = None
+
+
+def init(app):
+    global db
+    db = SQLAlchemy()
     db.init_app(app)
     # with app.app_context():
     #     # Extensions like Flask-SQLAlchemy now know what the "current" app